fix(route) : page 파일 구조 변경 (#52)
