+++ apps/admin/src/pages/common/NotFound.tsx
@@ -1,16 +1,1 @@
-import { useNavigate } from 'react-router-dom';
-import { FlexBox } from '@dudoong/ui';
-import { Button } from '@dudoong/ui';
-
-const NotFound = () => {
-  const navigate = useNavigate();
-  return (
-    <FlexBox align={'center'} justify={'center'}>
-      <Button onClick={() => navigate('/')} isLoading={false}>
-        잘못된 접근입니다.
-      </Button>
-    </FlexBox>
-  );
-};
-
-export default NotFound;
+export { default } from '../../../src/components/common/NotFound';
