+++ apps/admin/src/pages/events/Dashboard.tsx
@@ -1,13 +1,1 @@
-import { Button } from '@dudoong/ui';
-import { useNavigate } from 'react-router-dom';
-
-const Dashboard = () => {
-  const navigate = useNavigate();
-  return (
-    <>
-      <Button>어드민 이벤트 dashboard</Button>
-      <Button onClick={() => navigate('/')}>어드민 랜딩</Button>
-    </>
-  );
-};
-export default Dashboard;
+export { default } from '../../../src/components/events/Dashboard';
