+++ apps/admin/src/pages/events/Qr.tsx
@@ -1,13 +1,1 @@
-import { Button } from '@dudoong/ui';
-import { useNavigate } from 'react-router-dom';
-
-const Qr = () => {
-  const navigate = useNavigate();
-  return (
-    <>
-      <Button>어드민 이벤트 qr</Button>
-      <Button onClick={() => navigate('/')}>어드민 랜딩</Button>
-    </>
-  );
-};
-export default Qr;
+export { default } from '../../../src/components/events/Qr';
