+++ apps/admin/src/pages/hosts/Info.tsx
@@ -1,13 +1,1 @@
-import { Button } from '@dudoong/ui';
-import { useNavigate } from 'react-router-dom';
-
-const Info = () => {
-  const navigate = useNavigate();
-  return (
-    <>
-      <Button>어드민 호스트 info</Button>
-      <Button onClick={() => navigate('/')}>어드민 랜딩</Button>
-    </>
-  );
-};
-export default Info;
+export { default } from '../../../src/components/hosts/Info';
